Implement seen-mails support
This is necessary so we do not re-download already-seen mails. Since this
tool is meant for archiving/backup'ing the mailboxes, we need to keep track
of the mails we already downloaded to avoid re-downloading them.

The easiest way to do this, is by writing a simple text file with
uidvalidity+uid since the IMAP protocol guarantees those to be unique for
a given folder - forever.